pyproject.toml
[project]
name = "ai-agent-kb"
version = "1.0.0"
description = "基于 RAG 的 AI 知识库问答系统"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "AI Agent Team" }
]
keywords = ["ai", "rag", "knowledge-base", "chatbot", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "openai>=1.3.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "chromadb>=0.4.18",
    "sentence-transformers>=2.2.2",
    "pypdf>=3.17.1",
    "python-docx>=1.1.0",
    "markdown>=3.5.1",
    "beautifulsoup4>=4.12.2",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "google-generativeai>=0.8.5",
]

[project.scripts]
ai-agent = "run:main"
ai-agent-server = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0", 
    "httpx>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
